+++ ams/extension/__init__.py
@@ -1,0 +1,5 @@
+"""
+Extension module.
+"""
+
+from ams.extension import eva  # NOQA