+++ docs/source/release-notes.rst
@@ -9,6 +9,21 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 Pre-v1.0.0
 ==========
 
+v0.9.7 (2024-05-24)
+-------------------
+
+This patch release add the Roadmap section in the release notes, to list out some potential features.
+It also drafts the EV Aggregation model based on the state space modelg, but the finish date remains unknown.
+
+References:
+
+[1] J. Wang et al., "Electric Vehicles Charging Time Constrained Deliverable Provision of Secondary
+Frequency Regulation," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024.3356948.
+
+- Fix OTDF calculation
+- Add parameter `dtype='float64'` and `no_store=False` in `MatProcessor` PTDF, LODF, and OTDF calculation, to save memory
+- Add placeholder parameter `Bus.type`
+
 v0.9.6 (2024-04-21)
 -------------------
 
@@ -227,3 +242,42 @@ v0.4 (2023-01)
 -------------------
 
 This release outlines the package.
+
+Roadmap
+=======
+
+This section lists out some potential features that may be added in the future.
+Note that the proposed features are not guaranteed to be implemented and subject to change.
+
+Electric Vehicle for Grid Service
+------------------------------------------
+
+A charging-time-constrained EV aggregation based on the state-space model
+
+References:
+
+[1] J. Wang et al., "Electric Vehicles Charging Time Constrained Deliverable Provision of Secondary
+Frequency Regulation," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024.3356948.
+
+[2] M. Wang et al., "State Space Model of Aggregated Electric Vehicles for Frequency Regulation," in
+IEEE Transactions on Smart Grid, vol. 11, no. 2, pp. 981-994, March 2020, doi: 10.1109/TSG.2019.2929052.
+
+Distribution OPF
+--------------------------
+
+- Distribution networks OPF and its LMP
+- DG siting and sizing considering energy equity
+
+References:
+
+[1] H. Yuan, F. Li, Y. Wei and J. Zhu, "Novel Linearized Power Flow and Linearized OPF Models for
+Active Distribution Networks With Application in Distribution LMP," in IEEE Transactions on Smart Grid,
+vol. 9, no. 1, pp. 438-448, Jan. 2018, doi: 10.1109/TSG.2016.2594814.
+
+[2] C. Li, F. Li, S. Jiang, X. Wang and J. Wang, "Siting and Sizing of DG Units Considering Energy
+Equity: Model, Solution, and Guidelines," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024.3350914.
+
+Planning
+--------------------------
+
+- Transmission expansion planning